Move spoiler tracking logic to new plugin-wrapper component SpoilerSerloStaticRenderer
